refactor: merge StringExtensions.Async.cs into StringExtensions.cs
