--- src/Xbehave.2.Core/StringExtensions.Async.cs
@@ -1,122 +1,0 @@
-﻿// <copyright file="StringExtensions.Async.cs" company="xBehave.net contributors">
-//  Copyright (c) xBehave.net contributors. All rights reserved.
-// </copyright>
-
-namespace Xbehave
-{
-    using System;
-    using System.Diagnostics.CodeAnalysis;
-    using System.Threading.Tasks;
-    using Xbehave.Sdk;
-
-    /// <summary>
-    /// Provides access to step definition methods.
-    /// </summary>
-    public static partial class StringExtensions
-    {
-        /// <summary>
-        /// Defines a step in the current scenario.
-        /// </summary>
-        /// <param name="text">The step text.</param>
-        /// <param name="body">The action that will perform the step.</param>
-        /// <returns>
-        /// An instance of <see cref="IStepBuilder"/>.
-        /// </returns>
-        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "x", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "x", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
-        public static IStepBuilder x(this string text, Func<Task> body)
-        {
-            var stepDefinition = new StepDefinition
-            {
-                Text = text,
-                Body = body == null ? default(Func<IStepContext, Task>) : c => body(),
-            };
-
-            CurrentThread.StepDefinitions.Add(stepDefinition);
-            return stepDefinition;
-        }
-
-        /// <summary>
-        /// Defines a step in the current scenario.
-        /// </summary>
-        /// <param name="text">The step text.</param>
-        /// <param name="body">The action that will perform the step.</param>
-        /// <returns>
-        /// An instance of <see cref="IStepBuilder"/>.
-        /// </returns>
-        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "f", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "f", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
-        public static IStepBuilder f(this string text, Func<Task> body)
-        {
-            return text.x(body);
-        }
-
-        /// <summary>
-        /// Defines a step in the current scenario.
-        /// </summary>
-        /// <param name="text">The step text.</param>
-        /// <param name="body">The action that will perform the step.</param>
-        /// <returns>
-        /// An instance of <see cref="IStepBuilder"/>.
-        /// </returns>
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "_", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.Naming", "CA1707:IdentifiersShouldNotContainUnderscores", Justification = "Fluent API")]
-        [CLSCompliant(false)]
-        public static IStepBuilder _(this string text, Func<Task> body)
-        {
-            return text.x(body);
-        }
-
-        /// <summary>
-        /// Defines a step in the current scenario.
-        /// </summary>
-        /// <param name="text">The step text.</param>
-        /// <param name="body">The action that will perform the step.</param>
-        /// <returns>
-        /// An instance of <see cref="IStepBuilder"/>.
-        /// </returns>
-        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "x", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "x", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
-        public static IStepBuilder x(this string text, Func<IStepContext, Task> body)
-        {
-            var stepDefinition = new StepDefinition { Text = text, Body = body, };
-            CurrentThread.StepDefinitions.Add(stepDefinition);
-            return stepDefinition;
-        }
-
-        /// <summary>
-        /// Defines a step in the current scenario.
-        /// </summary>
-        /// <param name="text">The step text.</param>
-        /// <param name="body">The action that will perform the step.</param>
-        /// <returns>
-        /// An instance of <see cref="IStepBuilder"/>.
-        /// </returns>
-        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "f", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "f", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
-        public static IStepBuilder f(this string text, Func<IStepContext, Task> body)
-        {
-            return text.x(body);
-        }
-
-        /// <summary>
-        /// Defines a step in the current scenario.
-        /// </summary>
-        /// <param name="text">The step text.</param>
-        /// <param name="body">The action that will perform the step.</param>
-        /// <returns>
-        /// An instance of <see cref="IStepBuilder"/>.
-        /// </returns>
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "_", Justification = "Fluent API")]
-        [SuppressMessage("Microsoft.Naming", "CA1707:IdentifiersShouldNotContainUnderscores", Justification = "Fluent API")]
-        [CLSCompliant(false)]
-        public static IStepBuilder _(this string text, Func<IStepContext, Task> body)
-        {
-            return text.x(body);
-        }
-    }
-}
