+++ src/Xbehave.2.Core/StringExtensions.cs
@@ -12,7 +12,7 @@ namespace Xbehave
     /// <summary>
     /// Provides access to step definition methods.
     /// </summary>
-    public static partial class StringExtensions
+    public static class StringExtensions
     {
         /// <summary>
         /// Defines a step in the current scenario.
@@ -132,6 +132,111 @@ namespace Xbehave
         [SuppressMessage("Microsoft.Naming", "CA1707:IdentifiersShouldNotContainUnderscores", Justification = "Fluent API")]
         [CLSCompliant(false)]
         public static IStepBuilder _(this string text, Action<IStepContext> body)
+        {
+            return text.x(body);
+        }
+
+        /// <summary>
+        /// Defines a step in the current scenario.
+        /// </summary>
+        /// <param name="text">The step text.</param>
+        /// <param name="body">The action that will perform the step.</param>
+        /// <returns>
+        /// An instance of <see cref="IStepBuilder"/>.
+        /// </returns>
+        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "x", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "x", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
+        public static IStepBuilder x(this string text, Func<Task> body)
+        {
+            var stepDefinition = new StepDefinition
+            {
+                Text = text,
+                Body = body == null ? default(Func<IStepContext, Task>) : c => body(),
+            };
+
+            CurrentThread.StepDefinitions.Add(stepDefinition);
+            return stepDefinition;
+        }
+
+        /// <summary>
+        /// Defines a step in the current scenario.
+        /// </summary>
+        /// <param name="text">The step text.</param>
+        /// <param name="body">The action that will perform the step.</param>
+        /// <returns>
+        /// An instance of <see cref="IStepBuilder"/>.
+        /// </returns>
+        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "f", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "f", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
+        public static IStepBuilder f(this string text, Func<Task> body)
+        {
+            return text.x(body);
+        }
+
+        /// <summary>
+        /// Defines a step in the current scenario.
+        /// </summary>
+        /// <param name="text">The step text.</param>
+        /// <param name="body">The action that will perform the step.</param>
+        /// <returns>
+        /// An instance of <see cref="IStepBuilder"/>.
+        /// </returns>
+        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "_", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.Naming", "CA1707:IdentifiersShouldNotContainUnderscores", Justification = "Fluent API")]
+        [CLSCompliant(false)]
+        public static IStepBuilder _(this string text, Func<Task> body)
+        {
+            return text.x(body);
+        }
+
+        /// <summary>
+        /// Defines a step in the current scenario.
+        /// </summary>
+        /// <param name="text">The step text.</param>
+        /// <param name="body">The action that will perform the step.</param>
+        /// <returns>
+        /// An instance of <see cref="IStepBuilder"/>.
+        /// </returns>
+        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "x", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "x", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
+        public static IStepBuilder x(this string text, Func<IStepContext, Task> body)
+        {
+            var stepDefinition = new StepDefinition { Text = text, Body = body, };
+            CurrentThread.StepDefinitions.Add(stepDefinition);
+            return stepDefinition;
+        }
+
+        /// <summary>
+        /// Defines a step in the current scenario.
+        /// </summary>
+        /// <param name="text">The step text.</param>
+        /// <param name="body">The action that will perform the step.</param>
+        /// <returns>
+        /// An instance of <see cref="IStepBuilder"/>.
+        /// </returns>
+        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "f", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "f", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.StyleCop.CSharp.NamingRules", "SA1300:ElementMustBeginWithUpperCaseLetter", Justification = "Fluent API")]
+        public static IStepBuilder f(this string text, Func<IStepContext, Task> body)
+        {
+            return text.x(body);
+        }
+
+        /// <summary>
+        /// Defines a step in the current scenario.
+        /// </summary>
+        /// <param name="text">The step text.</param>
+        /// <param name="body">The action that will perform the step.</param>
+        /// <returns>
+        /// An instance of <see cref="IStepBuilder"/>.
+        /// </returns>
+        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "_", Justification = "Fluent API")]
+        [SuppressMessage("Microsoft.Naming", "CA1707:IdentifiersShouldNotContainUnderscores", Justification = "Fluent API")]
+        [CLSCompliant(false)]
+        public static IStepBuilder _(this string text, Func<IStepContext, Task> body)
         {
             return text.x(body);
         }
